refactor GpuMemoryBufferHandle
(cherry picked from commit 0a4febda4b095f3ee42af618e4eab10e7f7f9e64)
