+++ Plugin~/WebRTCPlugin/GraphicsDevice/Cuda/CMakeLists.txt
@@ -3,4 +3,6 @@ target_sources(WebRTCLib
     CudaContext.cpp
     CudaContext.h
     ICudaDevice.h
+    CudaBufferHandle.cpp
+    CudaBufferHandle.h
 )
