+++ Plugin~/WebRTCPlugin/GraphicsDevice/D3D11/D3D11GraphicsDevice.cpp
@@ -8,8 +8,8 @@
 
 #include "D3D11Texture2D.h"
 #include "NvCodecUtils.h"
-#include "GpuMemoryBuffer.h"
 #include "GraphicsDevice/GraphicsUtility.h"
+#include "GraphicsDevice/Cuda/CudaBufferHandle.h"
 
 using namespace Microsoft::WRL;
 using namespace ::webrtc;
@@ -188,7 +188,7 @@ rtc::scoped_refptr<I420Buffer> D3D11GraphicsDevice::ConvertRGBToI420(ITexture2D*
         }
         cuCtxPopCurrent(NULL);
 
-        std::unique_ptr<GpuMemoryBufferHandle> handle = std::make_unique<GpuMemoryBufferHandle>();
+        std::unique_ptr<CudaBufferHandle> handle = std::make_unique<CudaBufferHandle>();
         handle->array = mappedArray;
         handle->resource = resource;
         return handle;
