+++ Plugin~/WebRTCPlugin/GraphicsDevice/D3D12/D3D12GraphicsDevice.cpp
@@ -7,6 +7,7 @@
 #include "D3D12Constants.h"
 #include "D3D12GraphicsDevice.h"
 #include "D3D12Texture2D.h"
+#include "GraphicsDevice/Cuda/CudaBufferHandle.h"
 #include "GraphicsDevice/D3D11/D3D11Texture2D.h"
 #include "GraphicsDevice/GraphicsUtility.h"
 #include "NvCodecUtils.h"
@@ -362,7 +363,7 @@ rtc::scoped_refptr<webrtc::I420Buffer> D3D12GraphicsDevice::ConvertRGBToI420(
         }
         cuCtxPopCurrent(NULL);
 
-        std::unique_ptr<GpuMemoryBufferHandle> handle = std::make_unique<GpuMemoryBufferHandle>();
+        std::unique_ptr<CudaBufferHandle> handle = std::make_unique<CudaBufferHandle>();
         handle->array = array;
         handle->externalMemory = externalMemory;
         return handle;
